chore(core): move create wallet step indicator v2
